Extract composables into files
