+++ app/src/main/kotlin/com/cristiangoncas/greenhousemonitor/ui/screen/heartbeat/HeartbeatScreen.kt
@@ -1,25 +1,22 @@
 package com.cristiangoncas.greenhousemonitor.ui.screen.heartbeat
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.cristiangoncas.greenhousemonitor.domain.entity.HeartBeat
-import com.cristiangoncas.greenhousemonitor.ui.screen.Screen
+import com.cristiangoncas.greenhousemonitor.ui.common.Screen
+import com.cristiangoncas.greenhousemonitor.ui.common.Loading
 
 @Composable
 fun HeartBeatScreen(viewModel: HeartbeatViewModel, innerPadding: PaddingValues) {
@@ -27,12 +24,7 @@ fun HeartBeatScreen(viewModel: HeartbeatViewModel, innerPadding: PaddingValues) 
 
     Screen {
         if (state.loading) {
-            Box(
-                modifier = Modifier.fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                CircularProgressIndicator()
-            }
+            Loading()
         } else {
             LazyColumn(
                 contentPadding = innerPadding,
