+++ app/src/main/kotlin/com/cristiangoncas/greenhousemonitor/ui/screen/home/HomeScreen.kt
@@ -1,18 +1,13 @@
 package com.cristiangoncas.greenhousemonitor.ui.screen.home
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -25,11 +20,9 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.cristiangoncas.greenhousemonitor.domain.entity.Average
-import com.cristiangoncas.greenhousemonitor.domain.entity.AverageTempHumid
-import com.cristiangoncas.greenhousemonitor.ui.screen.Screen
+import com.cristiangoncas.greenhousemonitor.ui.common.Screen
+import com.cristiangoncas.greenhousemonitor.ui.common.Loading
 
 @Composable
 fun HomeScreen(
@@ -80,31 +73,15 @@ fun HomeContent(
                 }
 
                 state.events?.let { events ->
-                    EventCountComponent(events.heaterOn.event, events.heaterOn.count, events.heaterOn.hours)
+                    EventCountComponent(
+                        events.heaterOn.event,
+                        events.heaterOn.count,
+                        events.heaterOn.hours
+                    )
                 }
             }
         }
     }
-}
-
-@Composable
-fun Loading() {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-    ) {
-        CircularProgressIndicator(
-            modifier = Modifier
-                .align(Alignment.Center)
-                .size(150.dp)
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun PreviewLoading() {
-    Loading()
 }
 
 @Composable
@@ -126,78 +103,6 @@ fun AverageDisplayComponent(label: String, value: String, modifier: Modifier = M
             textAlign = TextAlign.Center,
             fontSize = MaterialTheme.typography.titleLarge.fontSize,
             modifier = Modifier.weight(1f)
-        )
-    }
-}
-
-@Composable
-fun AveragesComponent(average: Average) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp), // Optional padding for the entire row
-        horizontalArrangement = Arrangement.SpaceAround // Space evenly between the components
-    ) {
-        AverageDisplayComponent(
-            label = "Avg Temperature ${average.hours}h",
-            value = average.avgTemp.toString()
-        )
-
-        AverageDisplayComponent(
-            label = "Avg Humidity ${average.hours}h",
-            value = average.avgHumid.toString()
-        )
-    }
-//    }
-}
-
-@Composable
-fun EventCountComponent(event: String, count: Int, hours: Int) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(100.dp)
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            text = "$event events - ${hours}h",
-            textAlign = TextAlign.Center,
-            fontSize = MaterialTheme.typography.titleMedium.fontSize,
-            modifier = Modifier.weight(1f)
-        )
-        Text(
-            text = count.toString(),
-            textAlign = TextAlign.Center,
-            fontSize = MaterialTheme.typography.titleLarge.fontSize,
-            modifier = Modifier.weight(1f)
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun EventCountComponentPreview() {
-    EventCountComponent(event = "Heater on", count = 2, hours = 24)
-}
-
-@Preview(showBackground = true)
-@Composable
-fun PreviewAverages() {
-    Column {
-        AveragesComponent(
-            Average(
-                avgTemp = 22f,
-                avgHumid = 78f,
-                hours = 6
-            )
-        )
-        AveragesComponent(
-            Average(
-                avgTemp = 22f,
-                avgHumid = 78f,
-                hours = 12
-            )
         )
     }
 }
