+++ app/src/main/kotlin/com/cristiangoncas/greenhousemonitor/ui/screen/logs/LogsScreen.kt
@@ -2,35 +2,23 @@ package com.cristiangoncas.greenhousemonitor.ui.screen.logs
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.DateRange
-import androidx.compose.material3.Card
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
 import androidx.compose.material3.pulltorefresh.pullToRefresh
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.cristiangoncas.greenhousemonitor.domain.entity.Event
 import com.cristiangoncas.greenhousemonitor.domain.entity.LogEntry
-import com.cristiangoncas.greenhousemonitor.ui.screen.Screen
+import com.cristiangoncas.greenhousemonitor.ui.common.Screen
+import com.cristiangoncas.greenhousemonitor.ui.common.Loading
 
 @Composable
 fun LogsScreen(
@@ -69,91 +57,6 @@ fun LogsContent(innerPadding: PaddingValues, state: LogsViewModel.UiState, onRef
             ) {
                 items(state.logs.size) { index ->
                     LogItem(state, index)
-                }
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun LogsScreenPreview() {
-    LogsScreen(
-        viewModel = viewModel(),
-        innerPadding = PaddingValues(0.dp)
-    )
-}
-
-@Composable
-fun Loading() {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-    ) {
-        CircularProgressIndicator(
-            modifier = Modifier
-                .align(Alignment.Center)
-                .size(150.dp)
-        )
-    }
-}
-
-@Composable
-fun LogItem(
-    state: LogsViewModel.UiState,
-    index: Int
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 28.dp, vertical = 4.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .padding(12.dp)
-        ) {
-            val log = state.logs[index]
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(6.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(
-                        imageVector = Icons.Filled.DateRange,
-                        contentDescription = Event.getEventByValue(log.event).label
-                    )
-                    Column(
-                        verticalArrangement = Arrangement.spacedBy(4.dp)
-                    ) {
-                        Text(
-                            modifier = Modifier,
-                            text = log.time
-                        )
-                        Text(
-                            modifier = Modifier,
-                            text = log.date
-                        )
-                    }
-                }
-            }
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(6.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Row(
-
-                ) {
-                    val event = Event.getEventByValue(log.event)
-                    Text(text = event.label + ": ")
-                    Text(text = log.data)
-
                 }
             }
         }
